feat: separate benchmark tests so they can be executed independently
